Refactor utility functions in the shared utils.js
