+++ tests/core-addon/integration/utils.js
@@ -5,6 +5,14 @@
 const { Builder, Capabilities } = require("selenium-webdriver");
 const firefox = require("selenium-webdriver/firefox");
 const fs = require("fs");
+const os = require("os");
+const path = require("path");
+
+// An object containing the sample remote-settings read from
+// a local file: they must contain, at least, the study template
+// add-on that has been published.
+const RALLY_TEST_STUDY_REGISTRY =
+  JSON.parse(fs.readFileSync("public/locally-available-studies.json"))[0];
 
 /**
  * Get a Selenium driver for using the Firefox browser.
@@ -57,4 +65,53 @@ async function getFirefoxDriver(headless, prefsOverride) {
     .build();
 }
 
-module.exports.getFirefoxDriver = getFirefoxDriver;
+/**
+ * Get a temporary directory.
+ *
+ * @returns {String} the path to a temporary directory.
+ */
+async function getTempDirectory() {
+  return await new Promise((resolve, reject) => fs.mkdtemp(
+      path.join(os.tmpdir(), 'rally-test-'),
+      (err, directory) => {
+        if (err) {
+          reject(err);
+        }
+        resolve(directory);
+      }
+    )
+  );
+}
+
+/**
+ * Initialize Remote Settings and populate the local database with test studies.
+ * Intended to be run by `driver.executeScript()` in CHROME context.
+ */
+async function initRemoteSettings(testStudy, timestamp) {
+  const { RemoteSettings } = ChromeUtils.import("resource://services-settings/remote-settings.js");
+  const remoteSettingsKey = "rally-studies-v1";
+
+  const db = await RemoteSettings(remoteSettingsKey).db;
+  await db.create(testStudy);
+  await db.importChanges({}, timestamp);
+
+  await RemoteSettings(remoteSettingsKey).get();
+}
+
+/**
+ * Trigger a Remote Settings update.
+ */
+async function updateRemoteSettings(modifiedTestStudy, timestamp) {
+  const { RemoteSettings } = ChromeUtils.import("resource://services-settings/remote-settings.js");
+  const remoteSettingsKey = "rally-studies-v1";
+
+  await RemoteSettings(remoteSettingsKey).emit("sync", { data: { current: [modifiedTestStudy] } });
+}
+
+module.exports = {
+  RALLY_TEST_STUDY_REGISTRY,
+  getFirefoxDriver,
+  getTempDirectory,
+  initRemoteSettings,
+  updateRemoteSettings,
+};
